livekit/livekit-voice-agent/pyproject.toml
[project]
name = "livekit-voice-agent"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "livekit-agents[google,silero,turn-detector]~=1.4",
    "livekit-plugins-cartesia>=1.4.3",
    "livekit-plugins-deepgram>=1.4.3",
    "livekit-plugins-noise-cancellation~=0.2",
    "openai>=2.24.0",
    "python-dotenv>=1.2.1",
]
